Add retry support in client
Adds support for configuring a retry provider in the client. The
provider is used in the Do function to retry any requests that return a
status code >= 500.

Suggestions for further improvements is to make it possible to set which
status codes are retried, and potentially which request methods are retried.